+++ app/src/test/java/fr/inria/verveine/extractor/java/VerveineJTest_Generics.java
@@ -405,35 +405,6 @@ public class VerveineJTest_Generics extends VerveineJTest_Basic {
         	);
         }
     }
-    
-    @Test
-    public void testConcretizePrimitiveArray(){
-    	
-    	Attribute attribute = detectFamixElement(Attribute.class, "attributeUsingPrimitiveArrayConcretization");
-    	
-    	Collection<TConcretization> concretizations = ((ParametricEntityTyping)attribute.getTyping()).getConcretizations();
-    	assertEquals(1, concretizations.size());
-    	
-    	TConcretization theConcretization = this.firstElt(concretizations);
-    	assertTrue(theConcretization.getTypeArgument() instanceof PrimitiveType);
-    	
-    	assertEquals("int", ((PrimitiveType) theConcretization.getTypeArgument()).getName());
-    }
-    
-    @Test
-    public void testConcretizePrimitiveArrayMixedWithRef(){
-    	
-    	Attribute attribute = detectFamixElement(Attribute.class, "attributeMixingPrimitiveArrayConcretizationWithRef");
-    	
-    	Collection<TConcretization> concretizations = ((ParametricEntityTyping)attribute.getTyping()).getConcretizations();
-    	assertEquals(2, concretizations.size());
-    	
-    	TConcretization theConcretization = this.elementAt(concretizations, 1);
-    	assertTrue(theConcretization.getTypeArgument() instanceof PrimitiveType);
-    	
-    	assertEquals("int", ((PrimitiveType) theConcretization.getTypeArgument()).getName());
- 
-    }
 
     // UTILITIES --------------------------------------------------
