+++ app/src/test/resources/generics/Nicolas.java
@@ -1,0 +1,6 @@
+package generics;
+
+public class Nicolas <T> {
+
+	Nicolas<Nicolas<Integer>> attr;
+}
