+++ app/src/test/resources/generics/UsingPrimitiveArrayConcretization.java
@@ -1,8 +1,8 @@
 package generics;
 
-public class UsingPrimitiveArrayConcretization {
+public class UsingPrimitiveArrayConcretization extends Nicolas<UsingPrimitiveArrayConcretization> {
 
-	private static final OneGenericClass<int[]> attributeUsingPrimitiveArrayConcretization = new OneGenericClass<int[]>();
+//	private static final OneGenericClass<int[]> attributeUsingPrimitiveArrayConcretization = new OneGenericClass<int[]>();
 	
-	private static final TwoGenericClass<UsingPrimitiveArrayConcretization, int[]> attributeMixingPrimitiveArrayConcretizationWithRef = new TwoGenericClass<UsingPrimitiveArrayConcretization, int[]>();
+//	private static final TwoGenericClass<UsingPrimitiveArrayConcretization, int[]> attributeMixingPrimitiveArrayConcretizationWithRef = new TwoGenericClass<UsingPrimitiveArrayConcretization, int[]>();
 }
